monster, player, IDamage, IHealing short > float으로 바꿈

+++ ProjectBS/Assets/_BsScripts/Effect/Effect.cs
@@ -48,7 +48,7 @@ public class Effect : MonoBehaviour, IPoolable
     {
         if (other.TryGetComponent(out IDamage<Monster> monster))
         {
-            monster.TakeDamage((short)Attack);
+            monster.TakeDamage(Attack);
         }
     }
