+++ ProjectBS/Assets/_BsScripts/MonsterScript/MonsterData/Base/MonsterData.cs
@@ -19,8 +19,8 @@ public abstract class MonsterData : ScriptableObject
     public float Ak => _attack;
     public float AkDelay => _attackDelay;
     public float Sp => _speed;
-    public short MaxHP => _maxHp;
-    public short Exp => _exp;
+    public float MaxHP => _maxHp;
+    public float Exp => _exp;
     public float Mass => _mass;
     public float Radius => _radius;
     public GameObject Prefab => _prefab;
@@ -33,8 +33,8 @@ public abstract class MonsterData : ScriptableObject
     [SerializeField] private float _attack;         // 몬스터 공격력
     [SerializeField] private float _attackDelay;    // 몬스터 공격딜레이
     [SerializeField] private float _speed;          // 몬스터 이동속도
-    [SerializeField] private short _maxHp;          // 몬스터 최대체력
-    [SerializeField] private short _exp;            // 몬스터 경험치
+    [SerializeField] private float _maxHp;          // 몬스터 최대체력
+    [SerializeField] private float _exp;            // 몬스터 경험치
     [SerializeField] private float _mass;           // 몬스터 무게
     [SerializeField] private float _radius;         // 몬스터 크기
     [SerializeField] private GameObject _prefab;    // 몬스터 프리팹
