+++ ProjectBS/Assets/_BsScripts/_Interface/IHealing.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHealing
 {
-    void ReceiveHeal(short dmg);
+    void ReceiveHeal(float dmg);
 }
